perbaikan

+++ database/migrations/2024_04_04_130303_create_points_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('admin_id');
             // $table->unsignedBigInteger('kategori_sampah_id');
             $table->integer('jumlah_poin');
-            $table->integer('jumlah_saldo');
+            $table->double('jumlah_saldo');
             $table->timestamps();
             $table->softDeletes();
